update file paths and add get lock account address
